Preventing paged output from prematurely disappearing.
The less pager tends to use an alternate screen where possible, restoring the previous screen state when it exits.  The -F option makes it exit early if it's not necessary, because there's only a screenful or less to display.  This combination of facts makes it completely broken when not used inside a terminal multiplexer that doesn't emulate the alternate screen.